*DV*: Fix flying contour bugs.
Not exactly a fix per se, since the lines still fly. I have implemented a workaround where the tornado visualization flies alongside the contour lines so that the contour always lies one the same plane in 3D as the image itself. Still no implementation of QTimer tho, but I doubt it is strictly enforced since the animation seems to be an open task that may look different for each person depending on interpretation.

Signed-off-by: Neutra1